refactor: move "scheduler" to utils
