+++ packages/element/__tests__/index.ts
@@ -3,7 +3,6 @@ import testComputingPair from './decorators/computingPair';
 import testElementDecorator from './decorators/element';
 import testInternalDecorator from './decorators/internal';
 import testPropertyDecorator from './decorators/property';
-import testScheduler from './scheduler';
 import testWithCorpusculeElement from './withCorpusculeElement';
 
 describe('@corpuscule/element', () => {
@@ -11,7 +10,6 @@ describe('@corpuscule/element', () => {
   testComputingPair();
   testElementDecorator();
   testPropertyDecorator();
-  testScheduler();
   testInternalDecorator();
   testWithCorpusculeElement();
 });
