+++ packages/utils/src/scheduler.d.ts
@@ -1,0 +1,3 @@
+declare const schedule: (callback: () => void) => Promise<void>;
+
+export default schedule;
